requirements-dev.txt
## Translation app
django-rosetta==0.6.8

## tests
selenium==2.24.0

## better python console
## cf. http://ipython.org/
#ipython==0.13

## ipython notebook dependencies
## requires libzmq-dev
#pyzmq==2.2.0
#tornado==2.3


## Werkzeug is needed to run django-admin.py runserver_plus (useful to debug)
## cf. http://werkzeug.pocoo.org/
Werkzeug==0.8.3

